refactor(k8s): move FakeRunner into a localcluster test file
